GameUtil to Game.proto
