+++ src/main.js
@@ -1,19 +1,20 @@
 var creepProto = require('Creep.proto');
 var roomProto  = require('Room.proto');
 var roomPositionProto = require('RoomPosition.proto');
+
 var memoryProto = require('Memory.proto');
+var gameProto   = require('Game.proto');
 
 var RMain = require('RMain');
-var GameUtil = require('GameUtil');
 
 module.exports.loop = function () {
 
     Object.assign(Creep.prototype, creepProto);
     Object.assign(Room.prototype, roomProto);
     Object.assign(RoomPosition.prototype, roomPositionProto);
-    Object.assign(Memory, memoryProto);
 
-    GameUtil.extend(Game);
+    Object.assign(Memory, memoryProto);
+    Object.assign(Game, gameProto);
 
     var rmain = new RMain({
       game: Game,
